backend/requirements.txt
fastapi
pymongo
python-socketio[client]
python-jose[cryptography]
uvicorn
python-dotenv
